Fix tests after rebasing
